storage: move `HealthStatus` into `types.rs`
It's a type, so it should be there.
